requirements.txt
astroid==2.6.2
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==2.2.19
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.10.1
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
